(MemAlloc) Refactor legacy code

+++ plugins/org.preesm.algorithm/src/org/preesm/algorithm/memory/allocation/DeGreefAllocator.java
@@ -109,7 +109,7 @@ public class DeGreefAllocator extends MemoryAllocator {
 
       // Alignment constraint
       long align = -1;
-      final Long typeSize = (Long) vertex.getPropertyBean().getValue(MemoryExclusionVertex.TYPE_SIZE);
+      final Long typeSize = vertex.getPropertyBean().getValue(MemoryExclusionVertex.TYPE_SIZE);
       if (this.alignment == 0) {
         align = typeSize;
       } else if (this.alignment > 0) {
