+++ plugins/org.preesm.algorithm/src/org/preesm/algorithm/memory/allocation/FirstFitAllocator.java
@@ -96,7 +96,7 @@ public class FirstFitAllocator extends OrderedAllocator {
       long freeFrom = 0; // Where the last exclusion ended
 
       long align = -1;
-      final Long typeSize = (Long) vertex.getPropertyBean().getValue(MemoryExclusionVertex.TYPE_SIZE);
+      final Long typeSize = vertex.getPropertyBean().getValue(MemoryExclusionVertex.TYPE_SIZE);
       // Alignment constraint
       if (this.alignment == 0) {
         align = typeSize;
